Security Reference and documentation polish
Adds a `acl` folder into the reference documentation, which will contain
grouped security references. This is based on the same process as the
config generation process and modified and extended the build script.

Updates security documentation graphics, which are not working on light
mode at the moment. These were contributed by @alve-crabnebula.

Polish the security documentation and mark most pages as no longer work
in progress.